+++ changelog.md
@@ -6,9 +6,12 @@ pre3:
 - Plugin can now register commands
 - !plugin list command
 - get plug by id
+pre4:
+- full version show at !version
+- dotfiles. separate plugin paths in autoplug file by newlines, btw.
+- just warn if file not found on !r
+- logger; logs into in default .frevi dir
 ## TODO
-- dotfiles
 - Editor events
 - Better plugin api overall...
-- Frevi dotfiles and plugin autoload
 - mb multithreading?